+++ src/PostgREST/Observation.hs
@@ -23,6 +23,7 @@ import Protolude hiding (toList)
 
 data Observation
   = AdminStartObs Text
+  | AdminServerCrashedObs SomeException
   | AppStartObs ByteString
   | AppServerAddressObs Text
   | ExitUnsupportedPgVersion PgVersion PgVersion